Add notice to discourage use
This role is very basic and was quickly put together adapted around an
existing configuration.

The 'mrlesmithjr.ipset' role looks to do everything this role does and
more in a much better way with more OS support.

https://galaxy.ansible.com/mrlesmithjr/ipset should be used.

This role will remain on GitHub for reference and for the deployments
that are still using it.